refactor: resolve CodeScene issues via policy module split
Extract alignment, contract, and source-parser helpers instead of suppressing code health rules for production policy tooling.
