+++ .deepsource.toml
@@ -9,6 +9,7 @@ exclude_patterns = [
   "scripts/lib/pr-ci-label-state.test.mjs",
   "scripts/validate-feature-exposure.mjs",
   "scripts/validate-feature-exposure.integration.test.mjs",
+  "scripts/lib/feature-exposure-source-parser.mjs",
   "scripts/lib/feature-exposure-policy.mjs",
   "server/**/*.test.js",
   "src/**/*.test.ts",
