+++ scripts/lib/feature-exposure-policy.mjs
@@ -6,6 +6,12 @@
  * and surface data from source files, then passes it here.
  */
 
+import {
+  validateClientServerExposureMatrices,
+  validateClientServerRegistryAlignment,
+} from './feature-exposure-policy-alignment.mjs';
+import { validateExposureTestContract } from './feature-exposure-policy-contract.mjs';
+
 /**
  * @typedef {{ ok: true }} PolicyOk
  * @typedef {{ ok: false, errors: string[] }} PolicyFail
@@ -25,13 +31,6 @@
 
 const BUILD_TARGET_LIST = ['local', 'beta', 'staging', 'production'];
 const ISO_DATE_PATTERN = /^\d{4}-\d{2}-\d{2}$/;
-
-const PER_FEATURE_TEST_PATTERNS = (featureKey) => [
-  `src/features/${featureKey}.test.tsx`,
-  `src/features/${featureKey}.test.ts`,
-  `src/features/${featureKey}/${featureKey}.test.tsx`,
-  `src/features/${featureKey}/${featureKey}.test.ts`,
-];
 
 /** Returns true when addedDate is a real YYYY-MM-DD calendar value. */
 function isValidIsoCalendarDate(addedDate) {
@@ -117,109 +116,6 @@ function validateServerCapabilities(registry, serverCapabilityKeys, errors) {
     if (!definition.serverCapabilityKey) continue;
     if (serverCapabilityKeys.includes(definition.serverCapabilityKey)) continue;
     errors.push(`Feature "${featureKey}" declares serverCapabilityKey "${definition.serverCapabilityKey}" but it is not in the server capability keys list.`);
-  }
-}
-
-/** Reports server registry keys that are absent from the client registry. */
-function validateServerFeaturesExistOnClient(registry, serverRegistry, errors) {
-  const clientKeys = new Set(Object.keys(registry));
-  for (const featureKey of Object.keys(serverRegistry)) {
-    if (clientKeys.has(featureKey)) continue;
-    errors.push(`Server registry feature "${featureKey}" is missing from client FEATURE_EXPOSURE_REGISTRY.`);
-  }
-}
-
-/** Reports client server-backed features missing from the server registry. */
-function validateServerBackedFeaturesExistOnServer(registry, serverRegistry, errors) {
-  const serverKeys = new Set(Object.keys(serverRegistry));
-  for (const [featureKey, definition] of Object.entries(registry)) {
-    if (!definition.serverBacked) continue;
-    if (serverKeys.has(featureKey)) continue;
-    errors.push(`Client server-backed feature "${featureKey}" is missing from SERVER_FEATURE_EXPOSURE_REGISTRY.`);
-  }
-}
-
-/** Reports orphan server capability keys with no matching client owner. */
-function validateServerCapabilitiesHaveClientOwners(registry, serverRegistry, errors) {
-  const clientCapabilityKeys = new Set(
-    Object.values(registry)
-      .filter((definition) => definition.serverBacked && definition.serverCapabilityKey)
-      .map((definition) => definition.serverCapabilityKey)
-  );
-
-  for (const [featureKey, definition] of Object.entries(serverRegistry)) {
-    if (clientCapabilityKeys.has(definition.serverCapabilityKey)) continue;
-    errors.push(
-      `Server capability key "${definition.serverCapabilityKey}" (feature "${featureKey}") has no matching client serverBacked feature.`
-    );
-  }
-}
-
-/** Aligns client and server registry keys and capability ownership. */
-function validateClientServerRegistryAlignment(registry, serverRegistry, errors) {
-  validateServerFeaturesExistOnClient(registry, serverRegistry, errors);
-  validateServerBackedFeaturesExistOnServer(registry, serverRegistry, errors);
-  validateServerCapabilitiesHaveClientOwners(registry, serverRegistry, errors);
-}
-
-/** Ensures client and server exposure matrices and capability keys stay aligned. */
-function validateClientServerExposureMatrices(registry, serverRegistry, errors) {
-  for (const [featureKey, serverDefinition] of Object.entries(serverRegistry)) {
-    const clientDefinition = registry[featureKey];
-    if (!clientDefinition) continue;
-
-    for (const target of BUILD_TARGET_LIST) {
-      const clientValue = clientDefinition.exposure?.[target];
-      const serverValue = serverDefinition.exposure?.[target];
-      if (clientValue !== serverValue) {
-        errors.push(
-          `Feature "${featureKey}" exposure.${target} is ${clientValue} on client but ${serverValue} on server registry.`
-        );
-      }
-    }
-
-    if (clientDefinition.serverCapabilityKey !== serverDefinition.serverCapabilityKey) {
-      errors.push(
-        `Feature "${featureKey}" serverCapabilityKey is "${clientDefinition.serverCapabilityKey}" on client but "${serverDefinition.serverCapabilityKey}" on server.`
-      );
-    }
-  }
-}
-
-/** Collects every feature key that is gated by routes, navigation, or side effects. */
-function collectGatedFeatures(surfaces, sideEffectModules) {
-  return new Set([
-    ...surfaces.gatedRoutes.map(({ feature }) => feature),
-    ...surfaces.navigationItems.map(({ feature }) => feature).filter(Boolean),
-    ...Object.keys(sideEffectModules),
-  ]);
-}
-
-/** Returns true when a gated feature has a per-feature test file on disk. */
-function hasPerFeatureTestFile(featureKey, existingTestFiles) {
-  return PER_FEATURE_TEST_PATTERNS(featureKey).some((pattern) => existingTestFiles.includes(pattern));
-}
-
-/** Returns true when a gated feature has a valid acknowledgement entry. */
-function hasValidAcknowledgement(featureKey, acknowledgements) {
-  const acknowledgement = acknowledgements[featureKey];
-  return (
-    typeof acknowledgement?.reason === 'string' &&
-    acknowledgement.reason.trim().length > 0 &&
-    typeof acknowledgement.trackingIssue === 'number' &&
-    acknowledgement.trackingIssue > 0
-  );
-}
-
-/** Requires exposure tests or explicit acknowledgement for every gated feature. */
-function validateExposureTestContract(contract, errors) {
-  const { surfaces, sideEffectModules, acknowledgements, existingTestFiles } = contract;
-  for (const featureKey of collectGatedFeatures(surfaces, sideEffectModules)) {
-    if (hasPerFeatureTestFile(featureKey, existingTestFiles)) continue;
-    if (hasValidAcknowledgement(featureKey, acknowledgements)) continue;
-    errors.push(
-      `Gated feature "${featureKey}" has no exposure test coverage or acknowledgement. Add a per-feature test or an entry to src/config/featureExposure.acknowledgements.json.`
-    );
   }
 }
